backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
chardet==4.0.0
dataclasses==0.6
deprecation==2.1.0
dj-database-url==2.3.0
django==5.1.4
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
django-htmx==1.21.0
django-ninja==1.3.0
django-taggit==6.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
gotrue==0.2.0
gunicorn==23.0.0
h11==0.14.0
httpcore==0.12.3
httpx==0.16.1
idna==2.10
iniconfig==2.0.0
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
postgrest-py==0.4.0
psycopg2-binary==2.9.10
py==1.11.0
pydantic==2.10.4
pydantic-core==2.27.2
pytest==6.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime-py==0.1.3
requests==2.25.1
rfc3986==1.5.0
ruff==0.8.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
supabase-py==0.0.2
toml==0.10.2
typing-extensions==4.12.2
urllib3==1.26.20
websockets==9.1
whitenoise==6.8.2
